ezkernel_debian.sh

+++ ezkernel_debian.sh
@@ -15,7 +15,7 @@ if [[ "$(id -u)" -ne 0 ]]; then
 fi
 
 fatal() {
-    printf '[FATAL] %s\n' "$*" >&2
+    printf '[FATAL] %s\n\n' "$*" >&2
     exit 1
 }
 
@@ -49,18 +49,18 @@ ART
         ;;
 esac
 
-printf "\n\nWelcome %s, to eZkernel for Debian.\n\n" "$USER"
-printf "The latest mainline Linux kernel from www.kernel.org will be compiled and installed.\n\n"
-printf "Checking kernels versions... please wait"
+printf '\n\n Welcome %s, to eZkernel for Debian.\n\n" "$USER'
+printf ' The latest mainline Linux kernel from www.kernel.org will be compiled and installed.\n\n'
+printf ' Checking kernels versions... please wait'
 apt-get update -qq || fatal " ERROR: no internet connection detected. Exiting."
 }
 apt-get install -y curl > /dev/null 2>&1
 
 KVER=$(curl -s https://www.kernel.org/finger_banner | sed -n '2s/^[^6]*//p')
 max_len=80
-printf "\r%-*s\n\n" "$max_len" "Checking kernels versions... done."
+printf '\r%-*s\n\n" "$max_len" "Checking kernels versions... done.'
 
-printf "Current kernel version: %s\nIt will be updated to:  %s\n\n" \
+printf ' Current kernel version: %s\nIt will be updated to:  %s\n\n' \
        "$(uname -r)" "$KVER"
 
 while true; do
@@ -77,7 +77,7 @@ while true; do
     fi
 done
 # user pressed Enter, continue.
-printf "\n\nChecking compilation dependencies...\n\n"
+printf '\n\n Checking compilation dependencies...\n\n'
 
 pkgs=(
     build-essential libdw-dev libelf-dev zlib1g-dev libncurses-dev
@@ -97,22 +97,21 @@ for p in "${pkgs[@]}"; do
     dpkg -s "$p" &>/dev/null || {
         apt-get install -y --no-install-recommends "$p" &>/dev/null && ((ok++))
     }
-    printf "\rProgress: %3d%% [%-20s] %-*s" \
+    printf "\rProgress: %3d%% [%-40s] %-*s" \
            $((i*100/sum)) \
-           "$(printf '|%.0s' $(seq 1 $((i*20/sum))))" \
+           "$(printf '|%.0s' $(seq 1 $((i*40/sum))))" \
            "$pkg_len" "$p"
 done
 
-printf "\rProgress: 100%% [%-20s] Installed %d new package(s).\n\n" \
-       "$(printf '|%.0s' $(seq 1 20))" "$ok"
+printf "\r Progress: 100%% [%-40s] Installed %d new package(s).\n\n" \
+       "$(printf '|%.0s' $(seq 1 40))" "$ok"
 
-printf "Downloading kernel sources...\n\n"
-
+printf ' Downloading kernel sources...\n\n'
 mkdir -p "kernel/linux-upstream-$KVER"
 cd "$_"
-
 wget https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/snapshot/linux-master.tar.gz
-printf 'Extracting kernel sources...\n\n'
+
+printf ' Extracting kernel sources...\n\n'
 tar -zxf *.gz --strip-components=1
 rm  *.gz
 
@@ -120,17 +119,30 @@ yes '' | make localmodconfig
 make menuconfig && (
     time make -j"$(nproc)" bindeb-pkg &&
     dpkg -i ~/kernel/*.deb &&
-    printf "\n\neZkernel compilation successful for version: %s\n" "$KVER"
+    printf "\n\n eZkernel compilation successful for version: %s\n" "$KVER"
 ) && reboot_system(){
-    printf "\nSystem will reboot now.\n\nPress Enter to continue or Ctrl+C to cancel"
-    read -rp '' && cd && rm -rf ~/kernel && {
+    printf '\n System will reboot now.\n\n'
+    while true; do
+    printf '\r\033[2K'
+    read -n1 -s -r -p ' Press Enter to continue or Ctrl+C to cancel.'
+    # check if User pressed Ctrl+C
+    if (( $? != 0 )); then
+        echo
+        exit 1
+    fi
+    # check if user pressed Enter (empty input)
+    if [[ -z "$REPLY" ]]; then
+        break
+    fi
+    done
+    cd && rm -rf ~/kernel && {
         sed -i 's/^GRUB_TIMEOUT=[0-9]\+/GRUB_TIMEOUT=1/' /etc/default/grub ||
         echo "GRUB_TIMEOUT=1" >> /etc/default/grub
     } && update-grub >/dev/null 2>&1 && reboot
 } && reboot_system || (
-    printf "\n\nCompilation or installation error. Exiting.\n\n"
-    exit 1
+    fatal " WARNING: compilation or installation error. Exiting.\n\n"
 )
+
